requirements.txt
Django == 3.1.4
psycopg2-binary == 2.8.6
djangorestframework == 3.12.2
Pillow == 8.0.1
django-filter == 2.4.0
mysqlclient
drf-yasg